+++ backend/src/main/java/com/example/griddominion/models/db/CoordinatesModel.java
@@ -6,35 +6,35 @@ import jakarta.persistence.Embeddable;
 @Embeddable
 public class CoordinatesModel {
     @Column(name = "x")
-    private int x;
+    private double x;
 
     @Column(name = "y")
-    private int y;
+    private double y;
 
 
 
     // Constructors
     public CoordinatesModel() {}
 
-    public CoordinatesModel(int x, int y) {
+    public CoordinatesModel(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
     // Getters and setters
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 }
